[Tests] Move test utils to a utils folder to remove it from code coverage
